+++ frontend/CHANGELOG.md
@@ -11,3 +11,4 @@
 * Create user using HTTP Request
 * Delete user using HTTP Request 
 * Upload file using VueJS, tiny-emitter and axios
+* Add markdown support for README.md